Add lh/rh splitting of non-recon-all cmds in plotly script
Previously, splitting cmds between lh and rh in plots was only done for
recon-all cmds.
Now, this is done for other freesurfer cmds, py scripts, etc. in which
hemis are specified.
Changes to some key functions, exemplary subject normalization, and
dataframe structuring in general were necessary to adapt to this.