Modify lh/rh cmd ordering strategy in get_top_x_cmds()
The ordering of lh+rh commands is now based on max(lh, rh) instead of mean(lh, rh).
To accomplish this, a reordering and subsequent iteration over df rows
is necessary. With this change, the function may take slightly longer to
execute (on test data: approx. 6 ms longer). This implementation may be improved in the future.