Fix issue of xlabels lingering after unselecting cmds in dash app
Occasionally, unselecting a cmd would remove the associated bar/box on
the plot, but the xlabel would remain.
This was due to plotly's categoryorder='mean ascending' functionality,
which was used to sort bars/boxes in ascending order of the means,
but introduced this as a side effect.
This was rectified by explicitly providing the desired order of x-axis
entries through the categoryarray argument.