Improve strategy for comparing to exemplary subject data
Instead of computing a ratio of the mean of a cmd of selected subjects
against the mean of the exemplary subject, the differences of the
selected subjects' cmd times to the mean cmd time of the exemplary
subject are now computed. The mean and variance of these differences are
then plotted. This also fixes an issue with displaying outlier subject
IDs after selecting an exemplary subject.