Switch to one-time yaml file loading in dash app, and adapt other files
This leads to faster dash app updating, since the previous
implementation involved reading all yaml files at every update.
Now, this is only done once in the beginning.
The downside is that, if any additional subject dirs are added to the
root dir after launching the app, they will not be read without
relaunching the app.